Adding bungee tab complete
